refactor(api): move SPDM trigger logic toward machine controller
Relocate SPDM attestation trigger helpers from the API attestation handler into
the machine state handler attestation module.

This keeps SPDM device scheduling logic with machine state handling and reduces
API handler coupling as part of the effort to move the machine state handler
into a separate crate.
