+++ crates/api/src/state_controller/machine/handler.rs
@@ -111,7 +111,7 @@ use crate::state_controller::machine::{
     MeasuringOutcome, get_measuring_prerequisites, handle_measuring_state,
 };
 
-mod attestation;
+pub mod attestation;
 mod dpf;
 mod helpers;
 mod machine_validation;
