+++ crates/api/src/state_controller/machine/handler/attestation.rs
@@ -18,11 +18,17 @@
 use std::sync::Arc;
 
 use carbide_redfish::libredfish::RedfishClientPool;
+use carbide_redfish::libredfish::error::state_handler_redfish_error as redfish_error;
 use carbide_uuid::machine::MachineId;
-use chrono::Utc;
+use chrono::{DateTime, Utc};
+use config_version::ConfigVersion;
+use itertools::Itertools;
+use libredfish::model::component_integrity::{ComponentIntegrities, ComponentIntegrity};
 use model::attestation::spdm::{
-    SpdmAttestationState, SpdmAttestationStatus, SpdmDeviceAttestationDetails, SpdmHandlerError,
+    SpdmAttestationState, SpdmAttestationStatus, SpdmDeviceAttestation,
+    SpdmDeviceAttestationDetails,
 };
+use model::bmc_info::BmcInfo;
 use model::machine::{
     AttestationMode, FailureCause, FailureDetails, FailureSource, MachineState, ManagedHostState,
     ManagedHostStateSnapshot, SpdmMeasuringState, StateMachineArea,
@@ -32,8 +38,161 @@ use state_controller::state_handler::{
     StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
 
-use crate::handlers::attestation as attestation_handlers;
 use crate::state_controller::machine::context::MachineStateHandlerContextObjects;
+
+pub async fn trigger_attestation(
+    db_pool: &PgPool,
+    redfish_client: Box<dyn libredfish::Redfish>,
+    bmc_info: &BmcInfo,
+    machine_id: &MachineId,
+    redfish_timeout_duration: std::time::Duration,
+) -> Result<u64, StateHandlerError> {
+    // retrieve bmc info for a machine and create redfish client
+    // get service root
+    // - absent -> return NotSupported
+    // get component integrities and create/insert device attestations
+    // - if none, return NotSupported
+
+    let service_root_future = redfish_client.get_service_root();
+
+    let service_root = match tokio::time::timeout(redfish_timeout_duration, service_root_future)
+        .await
+    {
+        Ok(redfish_result) => redfish_result.map_err(|e| redfish_error("get service root", e))?,
+        Err(_) => {
+            return Err(StateHandlerError::GenericError(eyre::eyre!(
+                "redfish service_root could not finish in {} secods",
+                redfish_timeout_duration.as_secs()
+            )));
+        }
+    };
+
+    if service_root.component_integrity.is_none() {
+        // let's treat 0 devices under attestation as NotSupported
+        return Ok(0);
+    }
+
+    let component_integrities_future = redfish_client.get_component_integrities();
+
+    let component_integrities =
+        match tokio::time::timeout(redfish_timeout_duration, component_integrities_future).await {
+            Ok(redfish_result) => {
+                redfish_result.map_err(|e| redfish_error("get component integrities", e))?
+            }
+            Err(_) => {
+                return Err(StateHandlerError::GenericError(eyre::eyre!(
+                    "redfish get_component_integrities could not finish in {} secods",
+                    redfish_timeout_duration.as_secs()
+                )));
+            }
+        };
+
+    let components = get_components_supporting_spdm(&component_integrities);
+
+    if components.is_empty() {
+        // let's treat 0 devices under attestation as NotSupported
+        return Ok(0);
+    }
+
+    // The validation that list is not changed is done by SKU validation. SKU
+    // validation checks that the device profile is not changed over time. If any
+    // device list is changed and SKU validation is passed, means SRE has approved the
+    // change request.
+    // Validating again is not needed.
+    // Remove existing device list and over-write with this list.
+    let time_now = Utc::now();
+    let device_attestations = components
+        .into_iter()
+        .map(|x| from_component_integrity(x.clone(), machine_id, &time_now, bmc_info))
+        .collect_vec();
+
+    let mut txn = db_pool.begin().await?;
+
+    let records_inserted = db::attestation::spdm::insert_device_attestations(
+        &mut txn,
+        machine_id,
+        device_attestations,
+    )
+    .await?;
+
+    txn.commit().await?;
+
+    tracing::info!(
+        "SPDM attestation commenced for machine {}, scheduled {} SPDM device attestations",
+        machine_id,
+        records_inserted
+    );
+
+    Ok(records_inserted)
+}
+
+// Rules:
+// ComponentIntegrityTypeVersion should be >= 1.1.0.
+// ComponentIntegrityType should be SPDM.
+// ComponentIntegrityEnabled should be true.
+// Once these all conditions are true, a device can be proceed with attestation.
+fn get_components_supporting_spdm(integrities: &ComponentIntegrities) -> Vec<&ComponentIntegrity> {
+    let supported_versions = ["1.1.0"]; // This can be configurable value.
+    let mut supported_components = vec![];
+
+    for component in &integrities.members {
+        if !component.component_integrity_enabled {
+            // Component Integrity is not enabled
+            continue;
+        }
+
+        if component.component_integrity_type != "SPDM" {
+            // Not SPDM, may be TPM.
+            continue;
+        }
+
+        if !supported_versions.contains(&component.component_integrity_type_version.as_str()) {
+            continue;
+        }
+
+        supported_components.push(component);
+    }
+
+    supported_components
+}
+
+fn from_component_integrity(
+    integrity: ComponentIntegrity,
+    machine_id: &MachineId,
+    time_now: &DateTime<Utc>,
+    bmc_info: &BmcInfo,
+) -> SpdmDeviceAttestation {
+    let ca_certificate_link = integrity
+        .spdm
+        .map(|x| x.identity_authentication)
+        .map(|x| x.responder_authentication.component_certificate)
+        .map(|x| x.odata_id);
+
+    let evidence_target =
+        if let Some(Some(data)) = integrity.actions.map(|x| x.get_signed_measurements) {
+            Some(data.target)
+        } else {
+            None
+        };
+
+    SpdmDeviceAttestation {
+        machine_id: *machine_id,
+        device_id: integrity.id,
+        nonce: uuid::Uuid::new_v4(),
+        bmc_info: bmc_info.clone(),
+        state: SpdmAttestationState::FetchMetadata,
+        state_version: ConfigVersion::initial(),
+        state_outcome: None,
+        metadata: None,
+        ca_certificate_link,
+        ca_certificate: None,
+        evidence_target,
+        evidence: None,
+        started_at: *time_now,
+        cancelled_at: None,
+        completed_at: None,
+    }
+}
 
 /// When SPDM attestation failed, check whether attestation was restarted (admin / status) or
 /// disabled in config; if so, transition back to the appropriate measuring state based on
@@ -99,15 +258,14 @@ pub(crate) async fn handle_spdm_trigger_state(
         .await
         .map_err(StateHandlerError::from)?;
 
-    let devices_scheduled = attestation_handlers::trigger_attestation(
+    let devices_scheduled = trigger_attestation(
         db_pool,
         redfish_client,
         &mh_snapshot.host_snapshot.bmc_info,
         host_machine_id,
         std::time::Duration::MAX,
     )
-    .await
-    .map_err(|e| SpdmHandlerError::TriggerMeasurementFail(e.to_string()))?;
+    .await?;
 
     // if 0 devices scheduled - this means it is unsupported
     // so we just proceed to the next state
